+++ draw.cpp
@@ -37,9 +37,9 @@ void initGfx() {
   loadTextures();
   
   // make a cursor
-  size_t pixels[15][15];
+  uint32_t pixels[15][15];
   memset(pixels, 0x00, sizeof(pixels));
-  for(size_t i = 0; i<15; i++){
+  for(uint32_t i = 0; i<15; i++){
     pixels[0][i]  = 0xffffffff;        // vert1
     pixels[14][i] = 0xffffffff;        // vert2
     pixels[i][0]  = 0xffffffff;        // horizontal1
@@ -57,15 +57,15 @@ void initGfx() {
   // cursor done
 }
 
-void drawTree(sSquare* tree, const size_t W, const size_t H, const float dX, const float dY){
+void drawTree(sSquare* tree, const uint32_t W, const uint32_t H, const float dX, const float dY){
   if(tree == nullptr) return;
   // draw lines
-  const size_t WH=W*H;
+  const uint32_t WH=W*H;
   
   glBegin(GL_LINES);
   glColor3ub(255,255,0);
-  for(size_t x=0; x<W; x++){
-    for(size_t y=0; y<H; y++){
+  for(uint32_t x=0; x<W; x++){
+    for(uint32_t y=0; y<H; y++){
         // vertical
         glVertex2i(dX+x*GRID_SIZE,   dY+y*GRID_SIZE);
         glVertex2i(dX+x*GRID_SIZE,   dY+(y+1)*GRID_SIZE);
@@ -79,9 +79,9 @@ void drawTree(sSquare* tree, const size_t W, const size_t H, const float dX, con
   // data
   int dx = (int)dX;
   int dy = (int)dY;
-  for(size_t party=PA; party<PN; party++){
-    for(size_t x=0; x<W; x++){
-      for(size_t y=0; y<H; y++){
+  for(uint32_t party=PA; party<PN; party++){
+    for(uint32_t x=0; x<W; x++){
+      for(uint32_t y=0; y<H; y++){
         glColor3ub(party*255,255-party*255,0);
         drawInt(tree[party*WH+x*H+y].size,dx+x*GRID_SIZE,   dy+y*GRID_SIZE+party*15);
       }
@@ -98,7 +98,7 @@ void drawPlanets(const saPlanet & sPlanets, const float dX, const float dY){
   glBegin(GL_QUADS);
     const int R = PLANET_RADIUS; // radius
     
-    for(size_t i=0; i<sPlanets.size; i++){
+    for(uint32_t i=0; i<sPlanets.size; i++){
       if(planets[i].shieldActive) { // draw a shielded planet!
         glColor3ub(planets[i].party*150,150-planets[i].party*150,255);
       } else {
@@ -114,13 +114,13 @@ void drawPlanets(const saPlanet & sPlanets, const float dX, const float dY){
   // draw line to gathering point for created ships
   glBegin(GL_LINES);
     glColor3ub(100,100,100);
-    for(size_t i=0; i<sPlanets.size; i++){
+    for(uint32_t i=0; i<sPlanets.size; i++){
       glVertex2i(dX+planets[i].x ,dY+planets[i].y );
       glVertex2i(dX+planets[i].tx,dY+planets[i].ty);
     }
   glEnd();
   // draw Health and other numbers onto planet
-  for(size_t i=0; i<sPlanets.size; i++){
+  for(uint32_t i=0; i<sPlanets.size; i++){
     glColor3ub(250,250,25);
     drawInt(planets[i].party,     dX+planets[i].x,      dY+planets[i].y-R, 1);
     drawInt(planets[i].level[0],  dX+planets[i].x-R+20, dY+planets[i].y-R/2, 1);
@@ -140,9 +140,9 @@ void drawShips(const saShip * sShips, const float dX, const float dY){
   glEnable(GL_TEXTURE_2D);
   glBindTexture(GL_TEXTURE_2D, info.textures[TEX_SHIP].id);
   glBegin(GL_QUADS);
-    for(size_t party=0; party<PN; party++) {
+    for(uint32_t party=0; party<PN; party++) {
       sShip * ships = sShips[party].ships;
-      for(size_t i=0; i<sShips[party].size; i++){
+      for(uint32_t i=0; i<sShips[party].size; i++){
         if(ships[i].health>0){
           glColor3ub(party * 255, (1 - party) * 255, (10 - ships[i].health) * 25);
           glTexCoord2f(0.0,0.0); glVertex2f(dX+ships[i].x-SHIP_RADIUS,dY+ships[i].y-SHIP_RADIUS);
@@ -155,13 +155,13 @@ void drawShips(const saShip * sShips, const float dX, const float dY){
   glEnd();
   glDisable(GL_TEXTURE_2D); 
 }
-void drawShipMarkers(const sShip * const ships, std::vector<size_t> & selectedShips, const float dX, const float dY){
+void drawShipMarkers(const sShip * const ships, std::vector<uint32_t> & selectedShips, const float dX, const float dY){
   glEnable(GL_TEXTURE_2D);
   glBindTexture(GL_TEXTURE_2D, info.textures[TEX_SHIP_MARKER].id);
   glBegin(GL_QUADS);
   glColor3ub(255,255,255); // always white
   //glColor3ub(75, 255, 75);
-  for(size_t i : selectedShips){
+  for(uint32_t i : selectedShips){
     if(ships[i].health>0){
       //glColor3ub(party * 255, (1 - party) * 255, (10 - ships[i].health) * 25);
       glTexCoord2f(0.0,0.0); glVertex2f(dX+ships[i].x-SHIP_RADIUS,dY+ships[i].y-SHIP_RADIUS);
@@ -177,10 +177,10 @@ void drawShipMarkers(const sShip * const ships, std::vector<size_t> & selectedSh
 void drawShots(const saShot * sShots, const float dX, const float dY){
   //glDisable(GL_BLEND); // BLEND because of GL_LINE_SMOOTH, otherwise can be turned off
   glBegin(GL_LINES);
-    for(size_t party=0; party<PN; party++) {
+    for(uint32_t party=0; party<PN; party++) {
       glColor3ub(party*255,255-party*255,0);
       sShot * shots = sShots[party].shots;
-      for(size_t i=0; i<sShots[party].size; i++){
+      for(uint32_t i=0; i<sShots[party].size; i++){
         if(shots[i].timeToLive>0) {
           glVertex2f(dX+shots[i].x,dY+shots[i].y);
           glVertex2f(dX+shots[i].x+shots[i].dx*SHOT_LENGTH/SHOT_SPEED,dY+shots[i].y+shots[i].dy*SHOT_LENGTH/SHOT_SPEED);
@@ -243,14 +243,14 @@ void drawInt(int i, float strX, float strY, float stretchX, float stretchY){
     sprintf(s,"%i",i);
     drawString(s,strlen(s),strX,strY,stretchX, stretchY);
 }
-void drawString(const char* str, size_t strlen, float strX, float strY, float stretchXY){
+void drawString(const char* str, uint32_t strlen, float strX, float strY, float stretchXY){
     drawString(str,strlen,strX,strY,stretchXY,stretchXY);
 }
-void drawString(const char* str, size_t strlen, float strX, float strY, float stretchX, float stretchY){
+void drawString(const char* str, uint32_t strlen, float strX, float strY, float stretchX, float stretchY){
     glEnable(GL_TEXTURE_2D);
     glBindTexture(GL_TEXTURE_2D, info.textures[TEX_FONT].id);               // Select Our Texture
     glBegin(GL_QUADS);
-        for(size_t i=0;i<strlen;i++){
+        for(uint32_t i=0;i<strlen;i++){
             char c = str[i];
             float cx = ((c-' ')%16)/16.0f; // 16 chars per row, relative to width=1.0f
             float cy = ((c-' ')/16)/6.0f; // 16 chars per row, 6 per col, relative to height=1.0f
